refactor wip: move explorer bake functions to baker
